Write static metadata to video output
In this context, "static metadata" means "metadata that doesn't change
on a frame-by-frame basis". In other words, it's metadata that is
suitable for a video container.

Tested with regular video output, OpenGL playblasts and screencasts.

I've also updated metadata tooltip to show whether they're image-only
or image/video.

For some metadata it doesn't make sense to store it on the video
container level, as it may/will change for every frame. Such metadata
could be written to the video file as well, but would require encoding
into a subtitle track or something like that, so that it's
time-dependent. This is outside the scope of this patch.